# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Taskosaur API",
    "description": "API for Taskosaur - Task Management Application",
    "version": "1.0.0",
    "contact": {
      "email": "info@taskosaur.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/workspaces/{workspaceId}/members/{userId}": {
      "get": {
        "tags": ["roles"],
        "summary": "Get workspace member (Detailed)",
        "description": "Get details of a specific workspace member",
        "operationId": "getWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace member details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMember"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      },
      "put": {
        "tags": ["roles"],
        "summary": "Update workspace member (Detailed)",
        "description": "Update a workspace member's role",
        "operationId": "updateWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Member details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMember"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      },
      "delete": {
        "tags": ["roles"],
        "summary": "Remove workspace member (Detailed)",
        "description": "Remove a member from a workspace",
        "operationId": "removeWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workspace member removed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      }
    },
...
    "/projects/{projectId}/members/{userId}": {
      "get": {
        "tags": ["roles"],
        "summary": "Get project member (Detailed)",
        "description": "Get details of a specific project member",
        "operationId": "getProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project member details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMember"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      },
      "put": {
        "tags": ["roles"],
        "summary": "Update project member (Detailed)",
        "description": "Update a project member's role",
        "operationId": "updateProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Member details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMember"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      },
      "delete": {
        "tags": ["roles"],
        "summary": "Remove project member (Detailed)",
        "description": "Remove a member from a project",
        "operationId": "removeProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project member removed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      }
    },
...
    "/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "Get all projects",
        "description": "Returns all projects the authenticated user has access to, optionally filtered by workspace",
        "operationId": "getProjects",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "description": "Optional ID of the workspace to filter projects",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
           {
            "name": "workspaceSlug",
            "in": "query",
            "description": "Optional slug of the workspace to filter projects",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["projects"],
        "summary": "Create project",
        "description": "Create a new project in a workspace",
        "operationId": "createProject",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "description": "Project to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "409": {
            "description": "Slug already in use"
          }
        }
      }
    },
    "/projects/{slug}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get project by slug",
        "description": "Returns a single project identified by its slug",
        "operationId": "getProjectBySlug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project not found"
          }
        }
      },
      "put": {
        "tags": ["projects"],
        "summary": "Update project",
        "description": "Update an existing project",
        "operationId": "updateProject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Project details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project not found"
          },
          "409": {
            "description": "Slug already in use"
          }
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "Delete project",
        "description": "Delete a project",
        "operationId": "deleteProject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateProjectRequest": {
        "type": "object",
        "required": ["name", "workspaceId"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Q2 Campaign"
          },
          "description": {
            "type": "string",
            "example": "Marketing campaign for Q2"
          },
          "workspaceId": {
            "type": "string",
            "example": "ws-123"
          },
          "slug": {
            "type": "string",
            "example": "q2-campaign"
          }
        }
      },
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Updated Q2 Campaign"
          },
          "description": {
            "type": "string",
            "example": "Updated description for Q2 campaign"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "example": "https://example.com/project-avatar.png"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user-123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "example": "https://example.com/avatar.png"
          },
          "role": {
            "type": "string",
            "example": "User",
            "enum": ["Admin", "User"]
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": ["Active", "Inactive", "Pending"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-10T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "permission-123"
          },
          "name": {
            "type": "string",
            "example": "workspace:create"
          },
          "description": {
            "type": "string",
            "example": "Permission to create workspaces"
          },
          "resource": {
            "type": "string",
            "example": "workspace"
          },
          "action": {
            "type": "string",
            "example": "create"
          }
        }
      },
      "WorkspaceMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "workspace-member-123"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "workspace": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "workspace-123"
              },
              "name": {
                "type": "string",
                "example": "My Workspace"
              },
              "slug": {
                "type": "string",
                "example": "my-workspace"
              }
            }
          },
          "role": {
            "type": "string",
            "example": "Owner",
            "enum": ["Owner", "Admin", "Member", "Guest"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "UpdateWorkspaceMemberRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": {
            "type": "string",
            "example": "Admin",
            "enum": ["Admin", "Member", "Guest"]
          }
        }
      },
      "ProjectMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "project-member-123"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "project": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "project-123"
              },
              "name": {
                "type": "string",
                "example": "My Project"
              },
              "slug": {
                "type": "string",
                "example": "my-project"
              },
              "workspace": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "workspace-123"
                  },
                  "name": {
                    "type": "string",
                    "example": "My Workspace"
                  },
                  "slug": {
                    "type": "string",
                    "example": "my-workspace"
                  }
                }
              }
            }
          },
          "role": {
            "type": "string",
            "example": "Manager",
            "enum": ["Manager", "Member", "Viewer"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "UpdateProjectMemberRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": {
            "type": "string",
            "example": "Manager",
            "enum": ["Manager", "Member", "Viewer"]
          }
        }
      },
...
      "Project": {
        "type": "object",
        "required": ["id", "name", "slug", "workspace"],
        "properties": {
          "id": {
            "type": "string",
            "example": "proj-xyz"
          },
          "name": {
            "type": "string",
            "example": "Frontend Refactor"
          },
          "slug": {
            "type": "string",
            "example": "frontend-refactor"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "example": "https://example.com/project-avatar.png"
          },
          "workspace": {
            "type": "object",
            "required": ["id", "slug", "name"],
            "properties": {
               "id": {
                "type": "string",
                "example": "ws-abc"
              },
              "slug": {
                "type": "string",
                "example": "my-cool-workspace"
              },
              "name": {
                "type": "string",
                "example": "My Cool Workspace"
              }
            }
          }
        }
      },
...
    }
  }
}